created some more hooks
